+++ shared/core/data/src/commonMain/kotlin/com/yellowtrack/platform/core/data/DataModule.kt
@@ -6,6 +6,7 @@ import com.yellowtrack.platform.core.data.internal.SqlDelightClientRepository
 import com.yellowtrack.platform.core.data.internal.SqlDelightCodbRepository
 import com.yellowtrack.platform.core.data.internal.SqlDelightContractRepository
 import com.yellowtrack.platform.core.data.internal.SqlDelightCrewRepository
+import com.yellowtrack.platform.core.data.internal.SqlDelightDeliverableRepository
 import com.yellowtrack.platform.core.data.internal.SqlDelightExpenseRepository
 import com.yellowtrack.platform.core.data.internal.SqlDelightInvoiceRepository
 import com.yellowtrack.platform.core.data.internal.SqlDelightLeadRepository
@@ -40,6 +41,7 @@ val dataModule =
         single<CrewRepository> { SqlDelightCrewRepository(get(), get(), ioDispatcher) }
         single<TalentReleaseRepository> { SqlDelightTalentReleaseRepository(get(), get(), ioDispatcher) }
         single<PostProductionRepository> { SqlDelightPostProductionRepository(get(), get(), get(), ioDispatcher) }
+        single<DeliverableRepository> { SqlDelightDeliverableRepository(get(), get(), ioDispatcher) }
         single<ServiceTemplateRepository> { SqlDelightServiceTemplateRepository(get(), get(), get(), ioDispatcher) }
 
         single<LeadRepository> { SqlDelightLeadRepository(get(), get(), get(), ioDispatcher) }